fix: 修复Tensor计算库已知问题

+++ src/builtin/types/tensor.cpp
@@ -27,7 +27,7 @@ using namespace std;
 unordered_map<string, string> TensorType::staticMethods_;
 
 TensorType::TensorType(const vector<size_t> &shape)
-    : OtherType(typeCode()), shape_(shape), element_type_(Type::Double()) {
+    : OtherType(typeCode()), shape_(shape), element_type_(Type::Float()) {
     if (shape_.size() == 0) {
         throw invalid_argument("Tensor shape must at least have 1 dim");
     }
